Users/kzryzstof/day 10 mod int (#3)
* feat: enhance Gaussian elimination process with improved variable handling and logging, add voltage printing for machines

* feat: refactor matrix handling to use int64 for values, update related tests and logging

* feat: update Gaussian elimination to use Hermite Nominal Form, enhance pivoting logic, and adjust test case identifiers

* feat: improve row swapping logic in Gaussian elimination and add math import for enhanced calculations

* feat: improve row swapping logic in Gaussian elimination and add math import for enhanced calculations

* feat: refactor button handling and improve row reduction logic in Gaussian elimination

* feat: update Hermite Normal Form logic to improve variable handling and enhance row elimination process

* feat: add tests for Hermite Normal Form to validate unique solutions and free variable detection

* docs: update README for Day 10 problem description and solution approach

+++ day_10/internal/abstractions/counter.go
@@ -1,7 +1,7 @@
 package abstractions
 
 type Counter struct {
-	value uint32
+	value int64
 }
 
 func NewCounter() *Counter {
@@ -12,7 +12,7 @@ func (c *Counter) Increment() {
 	c.value++
 }
 
-func (c *Counter) GetValue() uint32 {
+func (c *Counter) GetValue() int64 {
 	return c.value
 }
